chore: move checks rules and endpoints to own dirs
