+++ ui/src/notifications/endpoints/components/AuthMethodTypeDropdown.tsx
