+++ ui/src/notifications/endpoints/components/EndpointOptions.tsx
