+++ ui/src/notifications/endpoints/components/EndpointOverlayFooter.tsx
