+++ ui/src/notifications/endpoints/components/EndpointTypeDropdown.tsx
