+++ ui/src/notifications/endpoints/components/MethodTypeDropdown.tsx
