+++ ui/src/resources/components/GetResources.tsx
@@ -5,12 +5,12 @@ import {connect} from 'react-redux'
 // Actions
 import {getAuthorizations} from 'src/authorizations/actions/thunks'
 import {getBuckets} from 'src/buckets/actions/thunks'
-import {getChecks} from 'src/alerting/actions/checks'
+import {getChecks} from 'src/checks/actions'
 import {getDashboards} from 'src/dashboards/actions/thunks'
-import {getEndpoints} from 'src/alerting/actions/notifications/endpoints'
+import {getEndpoints} from 'src/notifications/endpoints/actions'
 import {getLabels} from 'src/labels/actions'
 import {getMembers} from 'src/members/actions/thunks'
-import {getNotificationRules} from 'src/alerting/actions/notifications/rules'
+import {getNotificationRules} from 'src/notifications/rules/actions'
 import {getPlugins} from 'src/dataLoaders/actions/telegrafEditor'
 import {getScrapers} from 'src/scrapers/actions/thunks'
 import {getTasks} from 'src/tasks/actions/thunks'
